requirements.txt
streamlit>=1.30.0
pdfminer.six
python-docx
spacy>=3.8.2
sentence-transformers
openai
google-generativeai
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
scikit-learn
pandas
openpyxl
fpdf2
PyJWT>=2.8.0
bcrypt>=4.1.2
python-dotenv>=1.0.0
